Add pre-commit hooks for black and flake8
This commit adds a `.pre-commit-config.yaml` file that includes pre-commit hooks for the `black` and `flake8` tools. The `black` hook is set to version 22.12.0 from the `psf/black` repository, while the `flake8` hook is set to version 4.0.1 from the `pycqa/flake8` repository. Additionally, the `flake8` hook has an additional dependency on `pep8-naming`.